refactor: 아티스트 검색 로직 분리
